Relocated make offer logic to MyProductsViewModel for decreased complexity on ProductDetailViewModel.
