+++ src/commonMain/kotlin/onlytrade/app/viewmodel/product/di/MyProductsDI.kt
@@ -1,10 +1,13 @@
 package onlytrade.app.viewmodel.product.di
 
 import onlytrade.app.viewmodel.product.ui.MyProductsViewModel
+import onlytrade.app.viewmodel.product.ui.usecase.OfferUseCase
+import org.koin.core.module.dsl.factoryOf
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
 
 
 val myProductsModule = module {
     viewModelOf(::MyProductsViewModel)
+    factoryOf(::OfferUseCase)
 }
