+++ src/commonMain/kotlin/onlytrade/app/viewmodel/product/di/ProductDetailDI.kt
@@ -5,14 +5,12 @@ import onlytrade.app.viewmodel.product.offer.repository.data.remote.api.AddOffer
 import onlytrade.app.viewmodel.product.offer.repository.data.remote.api.DeleteOfferApi
 import onlytrade.app.viewmodel.product.offer.ui.usecase.WithdrawOfferUseCase
 import onlytrade.app.viewmodel.product.ui.ProductDetailViewModel
-import onlytrade.app.viewmodel.product.ui.usecase.OfferUseCase
 import org.koin.core.module.dsl.factoryOf
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
 
 val productDetailModule = module {
     viewModelOf(::ProductDetailViewModel)
-    factoryOf(::OfferUseCase)
     factoryOf(::WithdrawOfferUseCase)
     factoryOf(::OfferRepository)
     factoryOf(::AddOfferApi)
