+++ src/commonMain/kotlin/onlytrade/app/viewmodel/product/ui/ProductDetailViewModel.kt
@@ -12,14 +12,11 @@ import onlytrade.app.viewmodel.login.repository.LoginRepository
 import onlytrade.app.viewmodel.product.offer.repository.OfferRepository
 import onlytrade.app.viewmodel.product.offer.ui.usecase.WithdrawOfferUseCase
 import onlytrade.app.viewmodel.product.repository.data.db.Product
-import onlytrade.app.viewmodel.product.ui.nav.ProductDetailNav
 import onlytrade.app.viewmodel.product.ui.state.ProductDetailUiState
 import onlytrade.app.viewmodel.product.ui.state.ProductDetailUiState.GuestUser
 import onlytrade.app.viewmodel.product.ui.state.ProductDetailUiState.Idle
 import onlytrade.app.viewmodel.product.ui.state.ProductDetailUiState.LoadingOfferMade
 import onlytrade.app.viewmodel.product.ui.state.ProductDetailUiState.LoadingOfferReceived
-import onlytrade.app.viewmodel.product.ui.state.ProductDetailUiState.MakeOfferFail
-import onlytrade.app.viewmodel.product.ui.state.ProductDetailUiState.MakingOffer
 import onlytrade.app.viewmodel.product.ui.state.ProductDetailUiState.OfferDeleteApiError
 import onlytrade.app.viewmodel.product.ui.state.ProductDetailUiState.OfferMade
 import onlytrade.app.viewmodel.product.ui.state.ProductDetailUiState.OfferNotMade
@@ -27,12 +24,9 @@ import onlytrade.app.viewmodel.product.ui.state.ProductDetailUiState.OfferNotRec
 import onlytrade.app.viewmodel.product.ui.state.ProductDetailUiState.OfferReceived
 import onlytrade.app.viewmodel.product.ui.state.ProductDetailUiState.OfferRejected
 import onlytrade.app.viewmodel.product.ui.state.ProductDetailUiState.OfferWithdrawn
-import onlytrade.app.viewmodel.product.ui.state.ProductDetailUiState.OffersExceeded
 import onlytrade.app.viewmodel.product.ui.state.ProductDetailUiState.WithdrawingOffer
-import onlytrade.app.viewmodel.product.ui.usecase.OfferUseCase
 
 class ProductDetailViewModel(
-    private val offerUseCase: OfferUseCase,
     private val withdrawOfferUseCase: WithdrawOfferUseCase,
     private val loginRepository: LoginRepository,
     private val offerRepository: OfferRepository
@@ -41,21 +35,9 @@ class ProductDetailViewModel(
     var uiState: MutableStateFlow<ProductDetailUiState> = MutableStateFlow(Idle)
         private set
 
+
     private val user = loginRepository.user()
 
-    private lateinit var offeredProductIds: LinkedHashSet<Long>
-
-    init {
-        viewModelScope.launch {
-            ProductDetailNav.events.collect { event ->
-                when (event) {
-                    is ProductDetailNav.Event.TradeProducts -> {
-                        offeredProductIds = event.productIds
-                    }
-                }
-            }
-        }
-    }
 
     fun idle() {
         uiState.value = Idle
@@ -127,28 +109,6 @@ class ProductDetailViewModel(
             }.let { offer ->
                 uiState.value =
                     if (offer == null) OfferNotReceived else OfferReceived
-            }
-        }
-    }
-
-    fun makeOffer(productId: Long, offerReceiverId: Long) {
-        if (::offeredProductIds.isInitialized.not() || offeredProductIds.isEmpty())
-            return
-
-        uiState.value = MakingOffer
-
-        AppScope.launch {
-
-            uiState.value = when (val result = offerUseCase(
-                offerReceiverId = offerReceiverId,
-                offerReceiverProductId = productId,
-                offeredProductIds = offeredProductIds,
-            )) {
-                OfferUseCase.Result.OffersExceeded -> OffersExceeded
-                is OfferUseCase.Result.Error -> MakeOfferFail
-                is OfferUseCase.Result.OfferMade -> OfferMade(result.offer).apply {
-                    offeredProductIds.clear()
-                }
             }
         }
     }
