+++ src/commonMain/kotlin/onlytrade/app/viewmodel/product/ui/state/ProductDetailUiState.kt
@@ -11,9 +11,6 @@ sealed class ProductDetailUiState {
     data object OfferNotReceived : ProductDetailUiState()
     data class OfferMade(val offer: Offer) : ProductDetailUiState()
     data object OfferNotMade : ProductDetailUiState()
-    data object MakingOffer : ProductDetailUiState()
-    data object OffersExceeded : ProductDetailUiState()
-    data object MakeOfferFail : ProductDetailUiState()
     data object WithdrawingOffer : ProductDetailUiState()
     data object OfferRejected : ProductDetailUiState()
     data object OfferWithdrawn : ProductDetailUiState()
